R6 class ModelCompareMultivariate

R6 class ModelCompareMultivariate

R6 class ModelCompareMultivariate

Value

A dataframe containing the following columns 'Model': Name of the model 'Trend': The trend argument used in the VAR functions 'Season' The season argument used in the VAR functions 'SlidingASE': Whether Sliding ASE will be used for this model 'Init_K': The K value recommended by the VARselect function 'Final_K': The adjusted K value to take into account the smaller batch size (only when using sliding_ase)

Super class

tswgewrapped::ModelCompareBase -> ModelCompareMultivariateVAR

Methods

Public methods

ModelCompareMultivariateVAR$new()

ModelCompareMultivariateVAR$get_var_interest()

ModelCompareMultivariateVAR$get_data_var_interest()

ModelCompareMultivariateVAR$get_xIC()

ModelCompareMultivariateVAR$summarize_build()

ModelCompareMultivariateVAR$clone()

Inherited methods

tswgewrapped::ModelCompareBase$add_models()

tswgewrapped::ModelCompareBase$compute_metrics()

tswgewrapped::ModelCompareBase$get_batch_size()

tswgewrapped::ModelCompareBase$get_data()

tswgewrapped::ModelCompareBase$get_model_names()

tswgewrapped::ModelCompareBase$get_n.ahead()

tswgewrapped::ModelCompareBase$get_tabular_metrics()

tswgewrapped::ModelCompareBase$keep_models()

tswgewrapped::ModelCompareBase$plot_batch_ases()

tswgewrapped::ModelCompareBase$plot_batch_forecasts()

tswgewrapped::ModelCompareBase$plot_boxplot_ases()

tswgewrapped::ModelCompareBase$plot_simple_forecasts()

tswgewrapped::ModelCompareBase$remove_models()

tswgewrapped::ModelCompareBase$set_batch_size()

tswgewrapped::ModelCompareBase$set_verbose()

tswgewrapped::ModelCompareBase$statistical_compare()

Method new()

Initialize an object to compare several Univatiate Time Series Models

Usage

ModelCompareMultivariateVAR$new(
  data = NA,
  var_interest = NA,
  mdl_list,
  n.ahead = NA,
  batch_size = NA,
  step_n.ahead = TRUE,
  verbose = 0
)

Arguments

data

The dataframe containing the time series realizations (data should not contain time index)

var_interest

The output variable of interest (dependent variable)

mdl_list

A named list of all models (see format below)

n.ahead

The number of observations used to calculate ASE or forecast ahead

batch_size

If any of the models used sliding ase method, then this number indicates the batch size to use

step_n.ahead

If using sliding window, should batches be incremented by n.ahead (Default = TRUE)

verbose

How much to print during the model building and other processes (Default = 0)

Returns

A new `ModelCompareMultivariateVAR` object.

Method get_var_interest()

Returns the dependent variable name

Usage

ModelCompareMultivariateVAR$get_var_interest()

Returns

The dependent variable name

Method get_data_var_interest()

Returns the dependent variable data only

Usage

ModelCompareMultivariateVAR$get_data_var_interest()

Returns

The dependent variable data only

Method get_xIC()

Returns the AIC and the BIC for the model using the entire dataset

Usage

ModelCompareMultivariateVAR$get_xIC(sort_by = "AIC")

Arguments

sort_by

'AIC' or 'BIC'. Selects which column to sort the results by (Default: 'AIC')

Method summarize_build()

Returns the VAR model Build Summary

Usage

ModelCompareMultivariateVAR$summarize_build()

Method clone()

The objects of this class are cloneable with this method.

Usage

ModelCompareMultivariateVAR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.