Generate Multiple Realizations of a model Useful for checking model appropriateness

Generate Multiple Realizations of a model Useful for checking model appropriateness

generate_multiple_realization(
  x,
  phi = 0,
  theta = 0,
  d = 0,
  s = 0,
  vara = NA,
  n.realizations = 4,
  lag.max = 25,
  seed = NA,
  model_name = "Custom Model",
  return = "all"
)

Arguments

x
The actual time series realization

phi
'phi' values for the AR portion of the model (Default = 0)

theta
'theta' values for the AR portion of the model (Default = 0)

d
'd' value for the ARIMA portion of the model (Default = 0)

s
seasonality 's' of the ARIMA portion of the model (Default = 0)

vara
The white noise variance of the model

n.realizations
The number of hypothetical realizations to create for the model (Default = 4)

lag.max
The maximum amount of lag data to capture (Default = 25)

seed
The seed to use for reproducibility

model_name
The name to be assigned to the model (Default = "Custom Model")

return
What values to return (Default = 'all') 'all': Returns the actual along with the hypothetical realization data 'actual': Returns only the actual realization data 'hypothetical': Returns only the hypothetical realization data

Value

Named list with 'data' and 'results' 'data' contains the actual and the hypothetical realizations in tidy data format (appropriate for plotting) 'results' contain the ACF and Spectral Density values for the actual and hypothetical realizations in tidy data format

Examples

library(tswge)
data("sunspot.classic")
est = est.arma.wge(sunspot.classic,p=8)

#> 
#> Coefficients of Original polynomial:  
#> 1.2287 -0.4733 -0.1381 0.1569 -0.1403 0.0705 -0.1284 0.2069 
#> 
#> Factor                 Roots                Abs Recip    System Freq 
#> 1-1.5565B+0.8970B^2    0.8676+-0.6017i      0.9471       0.0965
#> 1-0.8771B              1.1402               0.8771       0.0000
#> 1-0.4147B+0.6550B^2    0.3166+-1.1944i      0.8093       0.2088
#> 1+0.7964B             -1.2557               0.7964       0.5000
#> 1+0.8231B+0.5043B^2   -0.8161+-1.1476i      0.7101       0.3484
#>   
#>   

r = generate_multiple_realization(x = sunspot.classic,
                                  phi = est$phi, theta = est$theta, vara = est$avar,
                                  seed = 11)